Add SpannableString.overlay and .overlayAll
The overlay methods will apply a provided span to matching
substrings of the receiver. In `.overlay`, it will apply to
only the first matching occurrence, while `.overlayAll` will
apply the span to all matching occurrences of the substring.

Because spans are not copyable, you provide a `spanBuilder` lambda
to generate a new span to apply to each occurrence.

This supports case-insensitive matching.