Add CharSequence.presence()
This is similar to the `String#presence` method in Ruby/Rails. Return
the string/charsequence if it has stuff in it, otherwise return null.

Useful for providing easy fallbacks when things might be null *or*
blank, and you want to have a non-blank string provided.

```kotlin
val name = getNameFromSomewhereItCouldBeAnything()
userName.text = name.presence() ?: "Unknown"
```